at4a35zar re06t6

+++ routes/web.php
@@ -3,14 +3,18 @@
 // ROTAS PARA TCCS
 //==============================
 
-Route::post('/create', 'WorkController@createWork');
-Route::get('/', 'WorkController@getWorks');
-Route::get('/form', 'WorkController@createForm');
+Route::post('/create', 'WorkController@createWork')->middleware('auth');
+Route::get('/', 'WorkController@shows');
+Route::get('/form', 'WorkController@createForm')->middleware('auth');
 Route::get('/show/{work}', 'WorkController@getWork');
-Route::get('/update/form/{work}', 'WorkController@updateForm');
-Route::put('/update/{work}', 'WorkController@updateWork');
-Route::delete('/delete/{work}', 'WorkController@deleteWork');
+Route::get('/update/form/{work}', 'WorkController@updateForm')->middleware('auth');
+Route::put('/update/{work}', 'WorkController@updateWork')->middleware('auth');
+Route::delete('/delete/{work}', 'WorkController@deleteWork')->middleware('auth');
 Route::get('/download/{filename}', 'WorkController@download');
-Route::get('/index2', 'WorkController@shows');
+Route::get('/admin', 'WorkController@getWorks')->middleware('auth');
 
 
+
+Auth::routes();
+
+Route::get('/home', 'HomeController@index')->name('home');
